backend/RAG/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1
langchain==0.3.19
langchain-community==0.3.19
langchain-groq==0.2.1
langchain-google-genai==2.0.8
langchain-huggingface==0.1.2
crewai==0.95.1
crewai-tools==0.20.1
sentence-transformers==3.4.1
faiss-cpu==1.9.0.post1
